Cleaned Up Type System
Previously the Type System was a concept in the internal packages. But
since it is required when building Tables, it was referenced in the public
API. This does not work. Hence we separated the Type's interfaces and
their implementation. The interfaces are public, of course, the
implementation is hidden in the internal parts.