Introduced daleq-spring
When we use Springs AbstractTransactional Tests we are bound to Spring's
Transaction handling especially Spring's way how to handle the connection
wrt to Transaction Management.

It is not possible just to work on the same jdbc Connection to obtain
the same transactional session (for obvious reason). Hence we have
to rely on Spring DataSourceUtils to obtain the connection within the
transaction. But this introduces a dependency to spring-jdbc, which we
do not want to enforce per se for daleq.

Hence there is daleq-core, with its Table DSL and then there is daleq-spring
with dependencies against Spring. This will open the door to add
further Spring helper classes in future.

As an example we've added an insert Test which would fail without the
changes mentioned above, because DaleqSupport would obtain a session which
is still waits for the other transaction to commit.