Rewrote DbUnit IDataSet generation
The initial implementation just created an XML file of the generated data.
This comes with the cost of a high memory footprint. The data is several
times in memory.

Now we have an adapter into the IDataSet world which just wraps TableData.
Therefore TableData and the data  it represents is only once in memory.
A further improvement might be having lazy TableData, but this is not yet
implemented.