Extracted FieldTypeReference to encapsulate FieldType resolution
We extracted the concept to resolve a FieldType in the RowBuilder. Now
it is not required, that actually a FieldDef is provided, but something
that can resolve a FieldType by a given TableType.
This approach would allow to implement other strategies, like
looking up by a name String or something.