Encapsulated TableType resolution
Like the FieldTypeReference, we've encapsulated the TableType as well.
Previously it was just possible to refer to a TableType with Daleq.table() by
a Class<T>. With this patch it is possible to refer to it with any
implementation of TableTypeReference. This would make an implementation
referencing by String possible.

Hence there can be future implementations, which construct a TableType
completely different than the current approach, like reading it from a
DB or something.