Introduced TableTypeRepository
Up to this commit, it was just possible to obtain a TableType only
by referencing a classes annotated with @TableDef. Now we decoupled
referring to a Table and obtaining a TableType.

TableTypes could be accessed via the TableTypeRepository, or more
in detail, a TableTypeResolver can be implemented to resolve a
TableType for a given TableTypeReference.

The default implementation behaves as before: The ClassBased*
implementation will resolve a table Type by the given class.

But now it is possible, to implement something like a
SQLTableTypeResolver, which will look into the DB to resolve the
actual TableType. Dynamic languages, here we come!