Changed Timestamp Template Creation
Previously we started at 1970-01-01 00:00:00. But Mysql fails with this value.
The time insertable into Mysql is 1970-01-01 01:00:00, which is hour new starting
point for Timestamp values.

Nontheless, it is allowed to change the behaviour of Template value, because this
is Daleq's internal behaviour. Unit Tests should not assume, how daleq will back
row identifiers to a specific value.