Integration tests can ignore a DB type
The @Restrict annotation can be applied to to Test Method. Currently
the rules requires a List of not supported DBs and a reason.
If this annotation is present, the test execution is skipped.

This enables us to write generic unit tests, but if a particular DB
cannot fullfill these generic behaviour, it could be left out in
such a test.