Removed already deprecated withRowsUntil.
The method was supposed to be used for convenience, but meanwhile
we learned that it is not intuitive. On one hand it is not a good idea to
start ids form 0, because some DBs treat 0 as an id special, especially
in combination with sequences. For instance, Mysql will not let you
explicitly set a a field, which has the auto increment attribute, with a
zero, only non zero values could be set explicitly. On the other hand,
specifying the maxId as an exclusive interval is not intuitive in this
context.

We deprecated it in v0.3.0 and therefore removed it in v0.4.0