Updated dependencies
The following compile dependencies have been updated:

* Spring: 3.1.1.RELEASE -> 3.2.0.RELEASE
* Guava: 12.0 -> 13.0.1
* commons-codec: 1.6 -> 1.7
* jodaTime: 1.6.2 -> 2.1

Furthermore these dependencies used in test scopes have been
updated:
* junit: 4.10 -> 4.11
* hamcrest: 1.3.RC2 -> 1.3
* HSQLDB: 2.2.8 -> 2.2.9
* H2: 1.3.167 -> 1.3.170
* MySQL: 5.1.21 -> 5.1.22

We missed to import jsr305 explicitly, although we use @Nullable.
This has been fixed.

There has been a bug in ImmutableTableData#getValuesOfField().
It did not threw if the given fieldName did not exist, but rather
returned a lazy evaluated Iterable, which in turn throw. The
unit test running with Junit 4.10 caught an exception as expected,
but only after checking on the iterable. With junit 4.11 this test
failed. Obviously is(nullValue()) works differently now. Anyways,
this has been fixed. The method correctly throws if the field
does not exist.

Clarified the usage of slf4j:
- We ain't use it explicitly, only DbUnit uses it.
- Since it is on the path, when something gets executed, we need one
  of its implementations, hence we opted for simple. Therefore we
  need the version DbUnit uses.
- Since we only require slf4j when we run the test, the testRuntime
  scope got that dependency.
- But Renamed dependency correctly to slf4jSimple

Updated the usage of SimpleJdbcTestUtils and SimpleJdbcTemplate to JdbcTestUtils and JdbcTemplate since the former where deprecated and solution has been provided with Spring 3.2